refactor: postlisttype명 ask로 수정
