+++ apps/pick-learn/src/views/PostDetail/ui/index.ts
@@ -1,0 +1,11 @@
+import BreadCrumbSection from './BreadCrumbSection';
+import PostDetailContentSection from './PostDetailContentSection';
+import PostDetailTitleSection from './PostDetailTitleSection';
+import PostDetailTopsection from './PostDetailTopsection';
+
+export {
+    BreadCrumbSection,
+    PostDetailContentSection,
+    PostDetailTitleSection,
+    PostDetailTopsection,
+};
